Update to qemu 4.0.0
You probably don't want to pull from my repository, but I'm not sure how to make a pull request to push that up to you. Feel free to just grab it somehow and add it to yours. My main goal here is to make you aware of the 4.0 ports' existence.